langchain-agents/pyproject.toml
[project]
name = "research-paper-agents"
version = "0.1.0"
description = "LangChain agents for research paper knowledge graph analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.0",
    "langchain-core>=0.1.0", 
    "langchain-openai>=0.1.0",
    "langchain-google-genai>=2.0.0",
    "google-generativeai>=0.8.0",
    "langserve[all]>=0.0.40",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "networkx>=3.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "plotly>=5.17.0"
]

[dependency-groups]
dev = [
    "langchain-cli>=0.0.37",
    "ruff>=0.1.0",
    "pytest>=7.4.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
